CleanUp 1. DataFiles Moved  to Folder 2. EntityAndDTO made to reference Class 3. Renamed DemoProviderAccessAPI to AzureSQL
