+++ DemoEntityAndDTO/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace DemoEntityAndDTO
+{
+    public class Class1
+    {
+    }
+}
