+++ DemoProviderAccessAPI/Properties/launchSettings.json
@@ -12,7 +12,7 @@
     "IIS Express": {
       "commandName": "IISExpress",
       "launchBrowser": true,
-      "launchUrl": "api/provider",
+      "launchUrl": "api/mintteaAZURESQLprovider",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
@@ -20,7 +20,7 @@
     "DemoProviderAccessAPI": {
       "commandName": "Project",
       "launchBrowser": true,
-      "launchUrl": "api/provider",
+      "launchUrl": "api/mintteaAZURESQLprovider",
       "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
