+++ DemoProviderSolution.sln
@@ -11,13 +11,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo_OnPremConsole_ToCosmos
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo_OnPremAPI_ToCosmos", "Demo_OnPremAPI_ToCosmos\Demo_OnPremAPI_ToCosmos.csproj", "{793F3833-B9D1-4C90-9157-E9A898C426C7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Demo_AzureAppService_ToCosmos", "Demo_AzureAppService_ToCosmos\Demo_AzureAppService_ToCosmos.csproj", "{FDBA3575-8B51-4645-A29F-12406A3193E2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo_AzureAppService_ToCosmos", "Demo_AzureAppService_ToCosmos\Demo_AzureAppService_ToCosmos.csproj", "{FDBA3575-8B51-4645-A29F-12406A3193E2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DataFiles_CosmosLoads", "DataFiles_CosmosLoads", "{585B653E-1799-49BF-9D5B-DBB3A8BE0D51}"
 	ProjectSection(SolutionItems) = preProject
 		OrangeBooster_Providers_1.csv = OrangeBooster_Providers_1.csv
 		WiredBrain_ProviderFile.csv = WiredBrain_ProviderFile.csv
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DemoEntityAndDTO", "DemoEntityAndDTO\DemoEntityAndDTO.csproj", "{6B6FF26F-145D-463A-8530-84DCC004A472}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -45,6 +47,10 @@ Global
 		{FDBA3575-8B51-4645-A29F-12406A3193E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FDBA3575-8B51-4645-A29F-12406A3193E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FDBA3575-8B51-4645-A29F-12406A3193E2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B6FF26F-145D-463A-8530-84DCC004A472}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B6FF26F-145D-463A-8530-84DCC004A472}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B6FF26F-145D-463A-8530-84DCC004A472}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B6FF26F-145D-463A-8530-84DCC004A472}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
